requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.5
pyyaml==6.0.2